+++ src/homebridge/platform.ts
@@ -44,7 +44,7 @@ export class HomebridgeDummyPlatform implements DynamicPlatformPlugin {
     this.Characteristic = api.hap.Characteristic;
 
     this.log = new Log(logger, config.verbose === true);
-    this.webhookManager = new WebhookManager(this.log, config.webhookPort);
+    this.webhookManager = new WebhookManager(this.log, this.api.user.configPath(), { port: config.webhookPort, ...config.webhookConfig });
     this.conditionManager = new ConditionManager(this.log, api.user.storagePath());
 
     this.log.always(